src/use-cases/search-gyms.ts
import type { Gym } from '@prisma/client'

import { GymsRepository } from '@/repositotories/gyms-repository'

interface SearchGymsUseCaseRequest {
  query: string
  page: number
}

interface SearchGymsUseUseCaseResponse {
  gyms: Gym[]
}

export class SearchGymsUseUseCase {
  constructor(private gymsRepository: GymsRepository) {}

  async execute({
    query,
    page,
  }: SearchGymsUseCaseRequest): Promise<SearchGymsUseUseCaseResponse> {
    const gyms = await this.gymsRepository.searchMany(query, page)

    return {
      gyms,
    }
  }
}
